feat: add iamauthz with auth-checking middleware
As a belt-and-suspenders approach to ensure that authorization has been
performed on all RPC methods in a service, this middleware requires
authorization implementations to call `iamauthz.Authorize(ctx)` to mark
the request as authorized.

Requests that return without being marked as authorized are stopped by
the middleware, instead returning INTERNAL (signaling that the
implementer of the RPC method has not correctly implemented
authorization).