feat: improve pattern resolution
This change updates the API of the IAM descriptor to use a per-service
custom struct with fields, and to unmarshal the descriptor fields from
precomputed raw proto messages resolved during compilation.

This enables resolution of resource patterns from other packages,
without having to import the package and resolve during runtime.