feat(iamspanner): provide functions in SetIamPolicy transaction
The added SetIamPolicyWithFunctionsInTransaction does the same thing as
SetIamPolicy used to do, but allows for passing in optional functions to
be called within the database transaction.

This allows for additional writes to be made within the same
transaction, making it possible e.g. to write to an outbox table within
the same transaction.